ItemFooterRow moved from SPC to FF
